change return type

+++ nativerl-policy/src/main/java/ai/skymind/nativerl/PolicyHelper.java
@@ -61,5 +61,5 @@ public interface PolicyHelper {
     /** Returns the discrete actions that should be performed in the given state. (Single Policy, Tuple Decisions) */
     long[] computeDiscreteAction(float[] state);
 
-    float[] computeActions(String url, String token, String postBody);
+    double[] computeActions(String url, String token, String postBody);
 }
